KACKEL FIle Transfer Server and Client
Software is Server / Client Software. Alowing user to send and recive fils over LAN or VPN

On Server side the you select file to send
The files is comprss on one side and uncompressed on the other
The Client recives the file and saves in seleted folder

ON both sides you will need to add Server ip Address of network card on PC used
and
Select a port on both sides. must be free and not blocked by fiawall and not being used

KACKLE.com.au

Coded by KACKLE